Build error in eclipse with chemsignal package
There was a build error in eclipse for the chemsignal package and its 
users, when creating a new project with the correct source paths.
this has been fixed by modifying the package references, and imports in
examples, for the chemsignal package. The package is now referred to as
'bsim.chemsignal', and imported as such.